Offload ethereum browser wallet signMessage to ethers.ts
